requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
pyyaml>=6.0

# Recommendation libraries
implicit>=0.7.0
lightfm>=1.17
surprise>=1.1.3
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Development and visualization
tqdm>=4.65.0
matplotlib>=3.7.0
plotly>=5.15.0
streamlit>=1.25.0
seaborn>=0.12.0

# Testing and code quality
pytest>=7.4.0
black>=23.7.0
ruff>=0.0.280
mypy>=1.5.0

# Optional: ML tracking
mlflow>=2.5.0
wandb>=0.15.0
